moved results (logic) to own file
