+++ src/client/src/components/stage/results/resultsTables/index.js
@@ -1,0 +1,1 @@
+export { default } from "./resultsTables"
